Refactor axios redirection to a kind of immutable version of it
